src/language experiments/requirements.txt
torch
datasets
transformers
tokenizers
huggingface_hub 
python-dotenv